// src/lib/cms/types.ts
export interface I18nText { fa: string; en: string; }
// ...
export interface Product {
  id: string;
  title: I18nText;
  description: I18nText;
  price: number;
  currency: 'IRR'|'USD'|'EUR'|'TMN'|'USDT';
  image?: string;
  fileUrl?: string;
  category: string;
  enabled: boolean;
  featured?: boolean;
  discount?: number;
  stock?: number;
}